Correctly report the required size to paint the PGroup control
Currently if the the content of the PGroup is smaller than the title
area width the text is cropped and possibly not readable because the
PGroup does not report the title area with as it preferred width.

This now changes the computation so that the hints are passed to the
render and then it could report the desired size.